+++ app/src/main/java/com/example/studyapp/MyAppCompose.kt
@@ -13,9 +13,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import androidx.navigation.compose.rememberNavController
 import com.example.studyapp.ui.constants.LEFT
 import com.example.studyapp.ui.constants.RIGHT
 import com.example.studyapp.ui.theme.StudyAppTheme
@@ -27,109 +26,108 @@ fun Preview() {
     //MyApp(navController = rememberNavController())
 }
 
-@Composable
-fun MyApp(navController: NavController,viewModel : QuestionsViewModel) {
-    StudyAppTheme {
-        // A surface container using the 'background' color from the theme
-        Surface(color = MaterialTheme.colors.background) {
-            Column(
-                modifier = Modifier
-                    .fillMaxHeight()
-                    .padding(8.dp)
-                    .padding(top = 60.dp),
-                verticalArrangement = Arrangement.Top,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Box(
-                    modifier = Modifier
-                        .height(200.dp)
-                        .fillMaxWidth(0.75f)
-                        .border(2.dp, Color.Black, RoundedCornerShape(15)),
-                    contentAlignment = Alignment.Center
-                ) {
-                    Text(
-                        text = "Android Quiz",
-                        modifier = Modifier
-                            .fillMaxWidth(),
-                        textAlign = TextAlign.Center
-                    )
-                }
-                Spacer(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(250.dp)
-                )
-                val modifier = Modifier.width(200.dp)
-                Row {
-                    ButtonColumn(
-                        start = 1,
-                        finish = 3,
-                        modifier = modifier,
-                        LEFT,
-                        navController,
-                        viewModel = viewModel
-                    )
-                    Divider(modifier = Modifier.width(3.dp))
-                    ButtonColumn(
-                        start = 4,
-                        finish = 6,
-                        modifier = modifier,
-                        id = RIGHT,
-                        navController,
-                        viewModel = viewModel
-                    )
-                }
-            }
-        }
-    }
-}
 
-
-@Composable
-fun ButtonColumn(
-    start: Int,
-    finish: Int,
-    modifier: Modifier,
-    id: String,
-    navController: NavController,
-    viewModel : QuestionsViewModel
-) {
-    Column(
-        verticalArrangement = Arrangement.Top,
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = modifier.layoutId(id)
-    ) {
-        for (i in start..finish) {
-            WeekButton(weekNumber = i, padding = 8.dp, modifier = Modifier.layoutId(i)) {
-                viewModel.currentWeek.postValue("Week $i")
-                navController.navigate("weekQuestions")
-            }
-            Divider(
-                modifier = modifier
-                    .width(150.dp)
-            )
-        }
-    }
-
-}
-
-@Composable
-fun WeekButton(weekNumber: Int, padding: Dp, modifier: Modifier, clickReaction: () -> Unit) {
-    Button(
-        onClick = {
-            clickReaction.invoke()
-        },
-        modifier = modifier
-            .width(150.dp)
-            .padding(padding)
-    ) {
-        Text(
-            text = "Week $weekNumber",
-            modifier = modifier
-                .fillMaxWidth()
-                .padding(padding),
-            textAlign = TextAlign.Center
-        )
-    }
-
-}
+//@Composable
+//fun MyApp(navController: NavController,questionsViewModel: QuestionsViewModel  = viewModel()) {
+//    StudyAppTheme {
+//        // A surface container using the 'background' color from the theme
+//        Surface(color = MaterialTheme.colors.background) {
+//            Column(
+//                modifier = Modifier
+//                    .fillMaxHeight()
+//                    .padding(8.dp)
+//                    .padding(top = 60.dp),
+//                verticalArrangement = Arrangement.Top,
+//                horizontalAlignment = Alignment.CenterHorizontally
+//            ) {
+//                Box(
+//                    modifier = Modifier
+//                        .height(200.dp)
+//                        .fillMaxWidth(0.75f)
+//                        .border(2.dp, Color.Black, RoundedCornerShape(15)),
+//                    contentAlignment = Alignment.Center
+//                ) {
+//                    Text(
+//                        text = "Android Quiz",
+//                        modifier = Modifier
+//                            .fillMaxWidth(),
+//                        textAlign = TextAlign.Center
+//                    )
+//                }
+//                Spacer(
+//                    modifier = Modifier
+//                        .fillMaxWidth()
+//                        .height(250.dp)
+//                )
+//                val modifier = Modifier.width(200.dp)
+//                Row {
+//                    ButtonColumn(
+//                        start = 1,
+//                        finish = 3,
+//                        modifier = modifier,
+//                        LEFT,
+//                        navController
+//                    )
+//                    Divider(modifier = Modifier.width(3.dp))
+//                    ButtonColumn(
+//                        start = 4,
+//                        finish = 6,
+//                        modifier = modifier,
+//                        id = RIGHT,
+//                        navController
+//                    )
+//                }
+//            }
+//        }
+//    }
+//}
+//
+//
+//@Composable
+//fun ButtonColumn(
+//    start: Int,
+//    finish: Int,
+//    modifier: Modifier,
+//    id: String,
+//    navController: NavController,
+//    questionsViewModel: QuestionsViewModel = viewModel()
+//) {
+//    Column(
+//        verticalArrangement = Arrangement.Top,
+//        horizontalAlignment = Alignment.CenterHorizontally,
+//        modifier = modifier.layoutId(id)
+//    ) {
+//        for (i in start..finish) {
+//            WeekButton(weekNumber = i, padding = 8.dp, modifier = Modifier.layoutId(i)) {
+//                questionsViewModel.currentWeek.postValue("Week $i")
+//                navController.navigate("weekQuestions")
+//            }
+//            Divider(
+//                modifier = modifier
+//                    .width(150.dp)
+//            )
+//        }
+//    }
+//
+//}
+//
+//@Composable
+//fun WeekButton(weekNumber: Int, padding: Dp, modifier: Modifier, clickReaction: () -> Unit) {
+//    Button(
+//        onClick = {
+//            clickReaction.invoke()
+//        },
+//        modifier = modifier
+//            .width(150.dp)
+//            .padding(padding)
+//    ) {
+//        Text(
+//            text = "Week $weekNumber",
+//            modifier = modifier
+//                .fillMaxWidth()
+//                .padding(padding),
+//            textAlign = TextAlign.Center
+//        )
+//    }
+//
+//}
