+++ app/src/main/java/com/example/studyapp/QuestionsListCompose.kt
@@ -12,6 +12,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Card
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.livedata.observeAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
@@ -24,6 +25,7 @@ import androidx.compose.ui.unit.sp
 import androidx.constraintlayout.compose.ConstraintLayout
 import androidx.constraintlayout.compose.ConstraintSet
 import androidx.constraintlayout.compose.Dimension
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import com.example.studyapp.model.Question
 import com.example.studyapp.ui.theme.Shapes
@@ -38,71 +40,74 @@ fun PreviewWeekList() {
 }
 
 
-@Composable
-fun WeekQuestions(navController : NavController,viewModel: QuestionsViewModel) {
-
-    val constraints = ConstraintSet {
-        val topText = createRefFor("weekText")
-        val listOfQuestions = createRefFor("questionList")
-
-        constrain(topText) {
-            top.linkTo(parent.top)
-            start.linkTo(parent.start)
-            end.linkTo(parent.end)
-            width = Dimension.wrapContent
-            height = Dimension.wrapContent
-        }
-
-        constrain(listOfQuestions) {
-            top.linkTo(topText.bottom)
-            start.linkTo(parent.start)
-            end.linkTo(parent.end)
-            bottom.linkTo(parent.bottom)
-            width = Dimension.fillToConstraints
-            height = Dimension.fillToConstraints
-        }
-    }
-    ConstraintLayout(constraintSet = constraints, Modifier.fillMaxSize()) {
-        viewModel.currentWeek.value?.let {
-            Text(
-                text = it,
-                textAlign = TextAlign.Center,
-                fontSize = 28.sp,
-                modifier = Modifier
-                    .layoutId("weekText")
-                    .padding(16.dp)
-            )
-        }
-        LazyColumn(
-            modifier = Modifier
-                .layoutId("questionList")
-                .fillMaxSize()
-        ) {
-            items(listQuesitons()) { question ->
-                val number = remember {
-                    mutableStateOf(question.id)
-                }
-                Card(
-                    elevation = 10.dp,
-                    shape = RoundedCornerShape(16.dp),
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp)
-                ) {
-                    Box(Modifier.clickable {
-                        viewModel.currentQuestion.postValue(question)
-                        navController.navigate("questionView")
-                    }.fillMaxSize()) {
-                        Text(
-                            text = "Question number ${number.value}", modifier = Modifier.padding(
-                                16.dp
-                            )
-                        )
-                    }
-
-                }
-            }
-        }
-    }
-}
-
+//@Composable
+//fun WeekQuestions(navController: NavController) {
+//
+//    val constraints = ConstraintSet {
+//        val topText = createRefFor("weekText")
+//        val listOfQuestions = createRefFor("questionList")
+//
+//        constrain(topText) {
+//            top.linkTo(parent.top)
+//            start.linkTo(parent.start)
+//            end.linkTo(parent.end)
+//            width = Dimension.wrapContent
+//            height = Dimension.wrapContent
+//        }
+//
+//        constrain(listOfQuestions) {
+//            top.linkTo(topText.bottom)
+//            start.linkTo(parent.start)
+//            end.linkTo(parent.end)
+//            bottom.linkTo(parent.bottom)
+//            width = Dimension.fillToConstraints
+//            height = Dimension.fillToConstraints
+//        }
+//    }
+//    ConstraintLayout(constraintSet = constraints, Modifier.fillMaxSize()) {
+//        questionsViewModel.currentWeek.value?.let {
+//            Text(
+//                text = it,
+//                textAlign = TextAlign.Center,
+//                fontSize = 28.sp,
+//                modifier = Modifier
+//                    .layoutId("weekText")
+//                    .padding(16.dp)
+//            )
+//        }
+//        LazyColumn(
+//            modifier = Modifier
+//                .layoutId("questionList")
+//                .fillMaxSize()
+//        ) {
+//            items(listQuesitons()) { question ->
+//                val number = remember {
+//                    mutableStateOf(question.id)
+//                }
+//                Card(
+//                    elevation = 10.dp,
+//                    shape = RoundedCornerShape(16.dp),
+//                    modifier = Modifier
+//                        .fillMaxWidth()
+//                        .padding(16.dp)
+//                ) {
+//                    Box(
+//                        Modifier
+//                            .clickable {
+//                                questionsViewModel.setCurrentQuestion(question = question)
+//                                navController.navigate("questionView")
+//                            }
+//                            .fillMaxSize()) {
+//                        Text(
+//                            text = "Question number ${number.value}", modifier = Modifier.padding(
+//                                16.dp
+//                            )
+//                        )
+//                    }
+//
+//                }
+//            }
+//        }
+//    }
+//}
+//
