+++ app/src/main/java/com/example/studyapp/repo/Repository.kt
@@ -1,0 +1,4 @@
+package com.example.studyapp.repo
+
+class Repository {
+}
