+++ app/src/main/java/com/example/studyapp/repo/RepositoryInterface.kt
@@ -1,0 +1,6 @@
+package com.example.studyapp.repo
+
+interface RepositoryInterface {
+    fun getQuestionsByWeek(week:String)
+
+}
